Create Beardo Prepare Recipe functions Major refactor - removed a lot of duplicate file reading code and replaced it with a javascript promise object.
