+++ package.json
@@ -29,6 +29,7 @@
     "argv": "0.0.2",
     "bootstrap": "^3.3.2",
     "cheerio": "*",
+    "es6-promise": "^2.0.1",
     "fs": "*",
     "inquirer": "0.8.0",
     "moment": "^2.9.0",
