+++ test/test.js
@@ -1,5 +1,5 @@
 var assert = require("assert");
-var Beardo = require("../Beardo");
+var Beardo = require("../beardo");
 var moment = require('moment');
 
 describe('Beardo', function(){
@@ -79,7 +79,9 @@ describe('Beardo', function(){
         '3', 'started tinkering', 'beardo', 'home', '0.00'],
       [ moment().add(8, 'days').format('YYYY-MM-DD'),
         '3', 'started tinkering', 'beardo', 'home', '0.00'],
-      [ moment().subtract(2, 'years').format('YYYY-MM-DD'),
+      [ moment().subtract(2, 'years')
+                .subtract(2, 'months')
+                .format('YYYY-MM-DD'),
         '3', 'started tinkering', 'beardo', 'home', '0.00']
       ];
